+++ app/src/main/java/com/inducesmile/temptoday/modals/singledayweathermodal/Wind.java
@@ -32,11 +32,11 @@ public class Wind implements Parcelable {
     private Double speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private Double deg;
 
     protected Wind(Parcel in) {
         this.speed = ((Double) in.readValue((Double.class.getClassLoader())));
-        this.deg = ((Integer) in.readValue((Integer.class.getClassLoader())));
+        this.deg = ((Double) in.readValue((Integer.class.getClassLoader())));
     }
 
     public Wind() {
@@ -50,11 +50,11 @@ public class Wind implements Parcelable {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
